Create image_training.py
Initially, I tried to train a machine learning algorithm to detect the LEGO pieces using YOLO, but decided against that as it would mean I would have to find a way to draw a box around the subject of each and every image I had generated. Instead, I decided to use PyTorch.